few utility changes
add extension parameter to make_output_filename so it can be used in
more situations.
expand the retry mechanism: auto detects a rate limiting message,
pausing for the amount of seconds that we are timed out; the max amount
of retries is now a parameter.
add defensive programming in case a uri_and_label only has a uri (checks
for '|' character)